feat: Refactoring of Breakdowns, and other changes (#371)

+++ builder/src/attribute/traits/get_bonuses.rs
@@ -1,3 +1,5 @@
+use rust_decimal::Decimal;
+
 use crate::bonus::Bonus;
 
 /// Implements the ability to get bonuses from different [`Attribute`] sub-types.
@@ -11,6 +13,7 @@ use crate::bonus::Bonus;
 ///
 /// ```
 /// use builder::{bonus::Bonus, attribute::GetBonuses};
+/// use rust_decimal::Decimal;
 ///
 /// enum Test {
 ///     A,
@@ -18,13 +21,13 @@ use crate::bonus::Bonus;
 /// }
 ///
 /// impl GetBonuses for Test {
-///     fn get_bonuses(&self, value: f32) -> Option<Vec<Bonus>> {
+///     fn get_bonuses(&self, value: Decimal) -> Option<Vec<Bonus>> {
 ///         // implementation
 ///         None // return None if there are none
 ///     }
 /// }
 ///
-/// let value = Test::A.get_bonuses(10f32);
+/// let value = Test::A.get_bonuses(Decimal::from(10));
 /// assert!(value.is_none());
 ///
 /// ```
@@ -33,6 +36,7 @@ use crate::bonus::Bonus;
 ///
 /// ```
 /// use builder::{bonus::Bonus, attribute::GetBonuses};
+/// use rust_decimal::Decimal;
 ///
 /// enum Test {
 ///     A,
@@ -43,19 +47,19 @@ use crate::bonus::Bonus;
 /// struct _TypeDEF;
 ///
 /// impl GetBonuses<_TypeABC> for Test {
-///     fn get_bonuses(&self, value: f32) -> Option<Vec<Bonus>> {
+///     fn get_bonuses(&self, value: Decimal) -> Option<Vec<Bonus>> {
 ///         None
 ///     }
 /// }
 ///
 /// impl GetBonuses<_TypeDEF> for Test {
-///     fn get_bonuses(&self, value: f32) -> Option<Vec<Bonus>> {
+///     fn get_bonuses(&self, value: Decimal) -> Option<Vec<Bonus>> {
 ///         Some(Vec::new())
 ///     }
 /// }
 ///
-/// let abc_value = GetBonuses::<_TypeABC>::get_bonuses(&Test::A, 10f32);
-/// let def_value = GetBonuses::<_TypeDEF>::get_bonuses(&Test::B, 10f32);
+/// let abc_value = GetBonuses::<_TypeABC>::get_bonuses(&Test::A, 10.into());
+/// let def_value = GetBonuses::<_TypeDEF>::get_bonuses(&Test::B, 10.into());
 ///
 /// assert!(abc_value.is_none());
 /// assert!(def_value.is_some());
@@ -75,5 +79,5 @@ pub trait GetBonuses<T = ()> {
     ///
     /// If there are no bonuses, for this object, this returns [`None`]. If there are bonuses, then
     /// a vector of each [`Bonus`] is returned.
-    fn get_bonuses(&self, value: f32) -> Option<Vec<Bonus>>;
+    fn get_bonuses(&self, value: Decimal) -> Option<Vec<Bonus>>;
 }
